feat: better setting layout ui
