+++ src/locales/en.json
@@ -277,7 +277,8 @@
   "Open UWP tool": "Open UWP tool",
   "Open UWP tool Info": "Since Windows 8, UWP apps (such as Microsoft Store) are restricted from directly accessing local host network services, and this tool can be used to bypass this restriction",
   "Update GeoData": "Update GeoData",
-  "Verge Setting": "Verge Setting",
+  "Verge Basic Setting": "Verge Basic Setting",
+  "Verge Advanced Setting": "Verge Advanced Setting",
   "Language": "Language",
   "Theme Mode": "Theme Mode",
   "theme.light": "Light",
