+++ src/pages/settings.tsx
@@ -11,7 +11,8 @@ import { useTranslation } from "react-i18next";
 import { BasePage, Notice } from "@/components/base";
 import { GitHub, HelpOutlineRounded, Telegram } from "@mui/icons-material";
 import { openWebUrl } from "@/services/cmds";
-import SettingVerge from "@/components/setting/setting-verge";
+import SettingVergeBasic from "@/components/setting/setting-verge-basic";
+import SettingVergeAdvanced from "@/components/setting/setting-verge-advanced";
 import SettingClash from "@/components/setting/setting-clash";
 import SettingSystem from "@/components/setting/setting-system";
 import { useThemeMode } from "@/services/states";
@@ -98,7 +99,17 @@ const SettingPage = () => {
               backgroundColor: isDark ? "#282a36" : "#ffffff",
             }}
           >
-            <SettingVerge onError={onError} />
+            <SettingVergeBasic onError={onError} />
+          </Box>
+        </Grid>
+        <Grid item xs={12} md={6}>
+          <Box
+            sx={{
+              borderRadius: 2,
+              backgroundColor: isDark ? "#282a36" : "#ffffff",
+            }}
+          >
+            <SettingVergeAdvanced onError={onError} />
           </Box>
         </Grid>
       </Grid>
